requirements.txt
flask
mysql-connector-python
bcrypt
jwt
requests